requirements.txt
mysql-connector-python==8.3.0
python-dotenv==1.0.1
pytz==2024.1
flask==2.3.3
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
pyserial==3.5
requests==2.31.0
PyJWT==2.8.0
RPi.GPIO